fix: change exit id length to 160-bits
Previously, we used 192-bits which may have overflown on exit priority.

Refers #283

+++ plasma_framework/contracts/mocks/exits/payment/routers/PaymentInFlightExitRouterMock.sol
@@ -37,11 +37,11 @@ contract PaymentInFlightExitRouterMock is PaymentInFlightExitRouter {
     }
 
     /** override and calls processInFlightExit for test */
-    function processExit(uint192 exitId, address ercContract) external {
+    function processExit(uint160 exitId, address ercContract) external {
         PaymentInFlightExitRouter.processInFlightExit(exitId, ercContract);
     }
 
-    function setInFlightExit(uint192 exitId, PaymentExitDataModel.InFlightExit memory exit) public {
+    function setInFlightExit(uint160 exitId, PaymentExitDataModel.InFlightExit memory exit) public {
         PaymentExitDataModel.InFlightExit storage ife = inFlightExitMap.exits[exitId];
         ife.isCanonical = exit.isCanonical;
         ife.exitStartTimestamp = exit.exitStartTimestamp;
@@ -60,15 +60,15 @@ contract PaymentInFlightExitRouterMock is PaymentInFlightExitRouter {
         }
     }
 
-    function getInFlightExitInput(uint192 exitId, uint16 inputIndex) public view returns (PaymentExitDataModel.WithdrawData memory) {
+    function getInFlightExitInput(uint160 exitId, uint16 inputIndex) public view returns (PaymentExitDataModel.WithdrawData memory) {
         return inFlightExitMap.exits[exitId].inputs[inputIndex];
     }
 
-    function setInFlightExitInputPiggybacked(uint192 exitId, uint16 inputIndex) public payable {
+    function setInFlightExitInputPiggybacked(uint160 exitId, uint16 inputIndex) public payable {
         inFlightExitMap.exits[exitId].setInputPiggybacked(inputIndex);
     }
 
-    function getInFlightExitOutput(uint192 exitId, uint16 outputIndex) public view returns (PaymentExitDataModel.WithdrawData memory) {
+    function getInFlightExitOutput(uint160 exitId, uint16 outputIndex) public view returns (PaymentExitDataModel.WithdrawData memory) {
         return inFlightExitMap.exits[exitId].outputs[outputIndex];
     }
